feat: buy now application data type
