+++ src/types/models/index.ts
@@ -142,25 +142,6 @@ export interface LeasingInterest {
   phone?: string
 }
 
-export interface MoneybackApplication {
-  email?: string
-  language?: string
-  firstName?: string
-  lastName?: string
-  birthdate: string
-  street: string
-  streetNumber: string
-  city: string
-  zipCode: string
-  serialNumber: string
-  gender: string
-  phone: string
-  contractStartDate: string
-  sellingPrice: number
-  vehicleConditionsConfirmation: boolean
-  termsAndConditionsConfirmation: boolean
-}
-
 export interface Date {
   month?: number
   year?: number
